WIP: MGMT-11455: Add formal API to allow users to set an additional trust bundle
Previously users had to use both the discovery ignition override and
install config override APIs to add trusted certs to the install
process. Discovery ignition overrides and install config overrides give
users a versatile way to customize the installation process, but at the
cost of "hackiness", instability and lack of support.

This commit adds a formal, single API for trusting certs - certs added
through this new API will automatically be included in both the
discovery ignition and the install config.

In the future a UI will be added to give users an easy way to use this
new API.